auth-svc/requirements.txt
bottle==0.12.13
pycrypto==2.6.1
python-jwt==2.0.1
